Move testPreloadedFeedsAreParsable to ViewDatabasePreloadTest
